refactor([DST-1419]): polish full-page data filter example
Polish pass on the canonical URL-state filter example so readers can
copy it without inheriting memo theatre, drift-prone types, or
oversized table components.

Structure:
- Rename to PascalCase, move hooks into hooks/ subfolder (one per
  concern: useFilter, useSearch, usePagination, useSort).
- Extract VenueRow, VenuesPaginationFooter, and CSV export into
  scannable pieces; VenuesTable's body now fits in a screen.

Source of truth:
- defaultFilter as concrete sentinels (0 / MAX_PRICE / []) feeds both
  parser defaults and active-filter detection (one rule). Removes
  per-field nullability branching and the duplicate VenueFilter type.
- setFilterFromForm encapsulates coercion + the slider's "max = no
  filter" rule; toolbar's onSubmit is three lines.

Hook composition:
- useSearch and useFilter call usePagination() directly to reset page
  (matches the production template).

A11y:
- Single live region on VisuallyHidden role="status"; drop the
  duplicate aria-live on the wrapper.
- Resolve "Venues" landmark duplication (Panel keeps it; wrapper is
  now a plain div).
- Honor prefers-reduced-motion on pagination smooth-scroll.

Other:
- Drop placeholder Add Venue and Delete buttons; strip internal
  DST-1288 references.
- Inverted traits check eliminates the readonly-tuple cast;
  remaining RAC typing-gap casts (Tag.Group.onRemove,
  Table.onSortChange) are documented inline.
