+++ docs/app/(examples)/examples/filter/page.tsx
@@ -1,32 +1,17 @@
 'use client';
 
-import {
-  Button,
-  Headline,
-  Inline,
-  Inset,
-  Panel,
-  Stack,
-  Text,
-} from '@marigold/components';
-import { Plus } from '@marigold/icons';
-import { AppliedFilter } from './applied-filter';
-import { Toolbar } from './toolbar';
-import { VenuesTable } from './venues-table';
+import { Headline, Inset, Panel, Stack, Text } from '@marigold/components';
+import { AppliedFilter } from './AppliedFilter';
+import { Toolbar } from './Toolbar';
+import { VenuesTable } from './VenuesTable';
 
 const FilterPage = () => (
   <Inset p={4}>
     <Stack space={8}>
-      <Inline alignX="between" alignY="top">
-        <Stack space={2}>
-          <Headline level={2}>Venues</Headline>
-          <Text>Browse and filter available venues for your events.</Text>
-        </Stack>
-        {/* Will be wired up in DST-1288 */}
-        <Button disabled>
-          <Plus /> Add Venue
-        </Button>
-      </Inline>
+      <Stack space={2}>
+        <Headline level={2}>Venues</Headline>
+        <Text>Browse and filter available venues for your events.</Text>
+      </Stack>
       <Panel aria-label="Venues">
         <Panel.Content>
           <Stack space="regular">
